+++ src/components/burger-constructor/index.js
@@ -1,15 +1,11 @@
 import styles from './burger-constructor.module.css'
-import scrollBarStyles from '../../helpers/scroll-bar.module.css'
-import {ConstructorElement, DragIcon} from "@ya.praktikum/react-developer-burger-ui-components";
+import {ConstructorElement} from "@ya.praktikum/react-developer-burger-ui-components";
 import {elementsShape, ingredientShape} from "../../shapes/shapes";
 import React, {useMemo} from "react";
 import Summary from "./summary";
 import PropTypes from "prop-types";
-import {useDispatch, useSelector} from "react-redux";
-import {arrayIsEmpty} from "../../helpers/collection-helper";
-import {removeIngredient} from "./actions/constructor-actions";
-import {useDrop} from "react-dnd";
-import {addIngredient} from "../burger-ingredients/actions/ingredients-actions";
+import {useSelector} from "react-redux";
+import FillersList from "./fillers-list";
 
 const BurgerConstructor = () => {
   const {bun, fillers} = useSelector(store => store.burgerConstructor);
@@ -61,56 +57,6 @@ const Bun = ({bun, type}) => {
   );
 }
 
-const FillersList = ({fillers}) => {
-  const dispatch = useDispatch();
-
-  const [{isHover}, dropRef] = useDrop({
-    accept: ['main', 'sauce'],
-    drop(ingredient) {
-      dispatch(addIngredient(ingredient));
-    },
-    collect: monitor => ({
-      isHover: monitor.isOver(),
-    })
-  });
-
-  if (arrayIsEmpty(fillers)) {
-    return (
-      // todo(kulikov): deal with margin and paddings
-      // todo(kulikov): deal with style
-      <div ref={dropRef} className={`${styles.emptyFiller} pr-4`} style={{backgroundColor: isHover ? 'pink' : '#37363F'}}>
-        Выберите начинку
-      </div>
-    );
-  }
-
-  const className = fillers.length > 5
-    ? `${styles.fillersList} ${scrollBarStyles.scrollBar} pr-2`
-    : `${styles.fillersList} pr-4`;
-
-  return (
-    <div ref={dropRef} className={className}>
-      {fillers.map(fillerData => <Filler key={fillerData.key} filler={fillerData}/>)}
-    </div>
-  );
-}
-
-const Filler = ({filler}) => {
-  const dispatch = useDispatch();
-
-  return (
-    <div className={styles.filler}>
-      <DragIcon type="primary"/>
-      <ConstructorElement
-        text={filler.name}
-        price={filler.price}
-        thumbnail={filler.image}
-        handleClose={() => dispatch(removeIngredient(filler._id, filler.key))}
-      />
-    </div>
-  );
-}
-
 Ingredients.propTypes = {
   bun: elementsShape,
   fillers: PropTypes.arrayOf(ingredientShape),
@@ -119,14 +65,6 @@ Ingredients.propTypes = {
 Bun.propTypes = {
   bun: ingredientShape,
   type: PropTypes.string.isRequired,
-};
-
-FillersList.propTypes = {
-  fillers: PropTypes.arrayOf(ingredientShape).isRequired,
-};
-
-Filler.propTypes = {
-  filler: ingredientShape.isRequired,
 };
 
 export default BurgerConstructor;
